pyproject.toml
[tool.poetry]
name = "pyreport"
version = "0.1.0"
description = "Simple framework for generating reports in python"
authors = ["PGrauman"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.2.2"
pypandoc = "^1.5"
atlassian-python-api = "*"
tabulate = "*"

[tool.poetry.dev-dependencies]
pytest = "^3.4"
pylint = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
